+++ CHANGELOG.md
@@ -23,3 +23,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Extract path roles map for Security Configuration
 - Project State Constraints Aspect
 - Frontend service
+- Set user as developer, pilot or final